Tests/LibWeb/Text/expected/interpolate-alpha-premultiplied.txt

oklab(0.627956 0.224863 0.125846 / 0.1)
